First implementation of file uploader. Uploaded content can also be viewed, and there is an extremely basic content browser.
Use const instead of var for stuff where possible in app.js
Add a configurable content directory, which is created if it does not exist, which is then served as a static content directory.
Add mutler to help with file uploading.
It uploads files, and you can customise the name if you want to. It will overwrite files that already exist with the same name.
Add filename sanitiser
Add unique filename generator module
First implementation of system to stop file name collsion when uploading
Started work on the content manager
Add configurable file size limit for uploads
Add extremely basic content browser

This will need some cleanup and polish later on, but for now it definitely works.